requirements.txt
openpyxl
python-dotenv
groq
pandas
streamlit
uvicorn